add and organize icons
